Introduce single `JavaObjectSerializationCodec`
Merging `SerializableWriteReplaceCodec` and `SerializableWriteObjectCodec`
together.
